requirements.txt
anthropic>=0.18.0
requests>=2.31.0
boto3>=1.28.0
google-api-python-client>=2.100.0
google-auth>=2.20.0
google-auth-oauthlib>=1.0.0